git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/linuxbuild_2017_06@6952 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ framework/I18N.cpp
@@ -359,7 +359,7 @@ int I18NLocal::LoadCharacterMapping( idStr& lang ) {
 		return 0;
 	}
 	
-	src.LoadMemory( buffer, strlen( buffer ), file );
+    src.LoadMemory(buffer, static_cast<int>(strlen(buffer)), file);
 	if ( !src.IsLoaded() ) {
 		common->Warning("I18N: Cannot load character remapping %s.", file.c_str() );
 		return 0;
