+++ game/ai/AAS_routing.cpp
@@ -165,7 +165,7 @@ void idAASLocal::CalculateAreaTravelTimes(void) {
 		}
 	}
 
-	assert( ( (unsigned int) bytePtr - (unsigned int) areaTravelTimes ) <= numAreaTravelTimes * sizeof( unsigned short ) );
+    assert(((ptrdiff_t)bytePtr - (ptrdiff_t)areaTravelTimes) <= static_cast<ptrdiff_t>(numAreaTravelTimes * sizeof(unsigned short)));
 }
 
 /*
