+++ game/anim/Anim_Blend.cpp
@@ -3802,7 +3802,7 @@ int idDeclModelDef::GetAnim( const char *name ) const {
 	int				numAnims;
 	int				len;
 
-	len = strlen( name );
+    len = static_cast<int>(strlen(name));
 	if ( len && idStr::CharIsNumeric( name[ len - 1 ] ) ) {
 		// find a specific animation
 		return GetSpecificAnim( name );
