+++ game/anim/Anim_Import.cpp
@@ -36,7 +36,7 @@ static idStr				Maya_Error;
 
 static exporterInterface_t	Maya_ConvertModel = NULL;
 static exporterShutdown_t	Maya_Shutdown = NULL;
-static int					importDLL = 0;
+static uintptr_t			importDLL = 0;
 
 bool idModelExport::initialized = false;
