+++ ui/GuiScript.h
@@ -49,7 +49,7 @@ public:
 	}
 	void FixupParms(idWindow *win);
 	size_t Size() {
-		int sz = sizeof(*this);
+		size_t sz = sizeof(*this);
 		for (int i = 0; i < parms.Num(); i++) {
 			sz += parms[i].var->Size();
 		}
@@ -79,7 +79,7 @@ public:
 		list.Append(gs);
 	}
 	size_t Size() {
-		int sz = sizeof(*this);
+		size_t sz = sizeof(*this);
 		for (int i = 0; i < list.Num(); i++) {
 			sz += list[i]->Size();
 		}
