+++ ui/SimpleWindow.h
@@ -40,7 +40,7 @@ public:
 	idStr			name;
 
 	idWinVar *		GetWinVarByName(const char *_name);
-	int				GetWinVarOffset( idWinVar *wv, drawWin_t* owner);
+    intptr_t		GetWinVarOffset(idWinVar *wv, drawWin_t* owner);
 	size_t			Size();
 
 	idWindow*		GetParent ( void ) { return mParent; }
